Rework Settings to house additional meta data (#133)

+++ cpp/src/qdk/chemistry/algorithms/microsoft/active_space/autocas_active_space.cpp
@@ -13,8 +13,8 @@ namespace qdk::chemistry::algorithms::microsoft {
 std::shared_ptr<data::Wavefunction> AutocasActiveSpaceSelector::_run_impl(
     std::shared_ptr<data::Wavefunction> wavefunction) const {
   // get settings
-  const size_t min_plateau_size = _settings->get<size_t>("min_plateau_size");
-  const size_t num_bins = _settings->get<size_t>("num_bins");
+  const int64_t min_plateau_size = _settings->get<int64_t>("min_plateau_size");
+  const int64_t num_bins = _settings->get<int64_t>("num_bins");
   const double entropy_threshold = _settings->get<double>("entropy_threshold");
   const bool normalize_entropies = _settings->get<bool>("normalize_entropies");
