+++ cpp/src/qdk/chemistry/algorithms/microsoft/active_space/valence_active_space.cpp
@@ -31,8 +31,9 @@ std::shared_ptr<data::Wavefunction> ValenceActiveSpaceSelector::_run_impl(
   }
 
   // Get the number of electrons and active orbitals from settings
-  int num_active_electrons = _settings->get<int>("num_active_electrons");
-  int num_active_orbitals = _settings->get<int>("num_active_orbitals");
+  int64_t num_active_electrons =
+      _settings->get<int64_t>("num_active_electrons");
+  int64_t num_active_orbitals = _settings->get<int64_t>("num_active_orbitals");
 
   // Validate settings
   if (num_active_electrons <= 0) {
