+++ cpp/src/qdk/chemistry/algorithms/microsoft/macis_base.cpp
@@ -22,7 +22,7 @@ macis::MCSCFSettings get_mcscf_settings_(const data::Settings& settings_) {
     SET_MACIS_SETTING(settings_, mcscf_settings, ci_max_subspace, size_t);
   } else {
     mcscf_settings.ci_max_subspace =
-        settings_.get<size_t>("davidson_iterations");
+        settings_.get<int64_t>("davidson_iterations");
   }
   SET_MACIS_SETTING(settings_, mcscf_settings, ci_matel_tol, double);
   return mcscf_settings;
